ObjectDetection/requirements.txt
torch
torchvision
transformers
datasets
timm
pycocotools
opencv-python
pillow
matplotlib
albumentations
tqdm
